feat(getProof): made a task
